♻️ カラーコード取得関数切り出し
